requirements.txt
asgiref==3.10.0
boolean.py==5.0
CacheControl==0.14.4
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
cyclonedx-python-lib==11.7.0
defusedxml==0.7.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.25.2
gunicorn==23.0.0
h11==0.16.0
idna==3.11
license-expression==30.4.4
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
packageurl-python==0.17.6
packaging==25.0
pillow==12.0.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
py-serializable==2.1.0
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
python-decouple==3.8
redis==7.1.0
requests==2.33.1
rich==14.3.3
sortedcontainers==2.4.0
sqlparse==0.5.5
tomli==2.4.1
tomli_w==1.2.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.6.0